Flask-web-app/requirements.txt
torch
opencv-python
numpy
Pillow
flask
werkzeug
requests
firebase_admin
pyrebase4